Fixing size_t printing error format.

+++ tools/auto_answer.c
@@ -189,8 +189,8 @@ int main(int argc, char *argv[]){
 		ms_usleep(50000);
 		if (print_stats) {
 			ms_message("*********************************");
-			ms_message("*Current number of calls   [%10i]  *",ms_list_size(linphone_core_get_calls(lc)));
-			ms_message("*Number of calls until now [%10i]  *",ms_list_size(linphone_core_get_call_logs(lc)));
+			ms_message("*Current number of calls   [%10u]  *",(unsigned int)ms_list_size(linphone_core_get_calls(lc)));
+			ms_message("*Number of calls until now [%10u]  *",(unsigned int)ms_list_size(linphone_core_get_call_logs(lc)));
 			ms_message("*********************************");
 			print_stats=FALSE;
 		}
